requirements.txt
python-telegram-bot[socks]==13.7
ephem==4.0.0.2